feat : midtrans notification

+++ database/migrations/2025_04_12_122507_create_plans_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('plans', function (Blueprint $table) {
             $table->id();
             $table->string('title');
-            $table->decimal('price', 10, 2);
+            $table->integer('price');
             $table->integer('duration');
             $table->enum('resolution', ['720p', '1080p', '2k']);
             $table->integer('max_devices');
